backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic